fix(NA): support bazel paths with whitespaces on windows
chore(NA): restore original bazel.go file

chore(NA): insert missing build tag to exclude unix on windows builds

docs(NA): missing comment
